requirements.txt
asgiref==3.8.1
Django==5.2.1
djangorestframework==3.16.0
dnspython==2.7.0
mongoengine==0.29.1
pymongo==4.13.0
python-decouple==3.8
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.9.0
